refactor : rename SocketController to SocketConnector
